refactor: Initialize Firebase Remote Config lazily and inject CrashlyticsLogger
This commit refactors the `RemoteConfigProviderImpl` on Android to initialize `FirebaseRemoteConfig` lazily. It also injects `CrashlyticsLogger` for logging errors during config updates.

**Key Changes:**

- **`RemoteConfigProvider.android.kt`:**
    - `FirebaseRemoteConfig` is now initialized within a `lazy` delegate.
    - Configuration settings (fetch interval based on `BuildConfig.isUAT`) and default values are set during this lazy initialization.
    - The `ConfigUpdateListener` now uses the injected `CrashlyticsLogger` to log errors instead of printing to the console.
    - `CrashlyticsLogger` is added as a constructor parameter to `RemoteConfigProviderImpl`.
- **`CommonModule.android.kt`:**
    - Removed the direct Koin definition for `FirebaseRemoteConfig`. This is now handled within `RemoteConfigProviderImpl`.
    - `RemoteConfigProviderImpl` instantiation in Koin now correctly passes the `CrashlyticsLogger` dependency.
